Move utils.py to scripts/
